+++ common/src/main/java/es/urjc/etsii/grafo/util/ArrayUtil.java
@@ -5,7 +5,10 @@ import es.urjc.etsii.grafo.util.random.RandomManager;
 /**
  * Util methods to manipulate collections and arrays that are not part of the standard java API
  */
+@SuppressWarnings("DuplicatedCode") // Due to primitive types some methods must be duplicated
 public class ArrayUtil {
+
+    private ArrayUtil(){}
 
     /**
      * Reverse an array
@@ -19,14 +22,14 @@ public class ArrayUtil {
     /**
      * Reverse a fragment inside an array from start to end (inclusive)
      *
-     * @param arr Array to reverse
+     * @param arr   Array to reverse
      * @param start start index, inclusive
      * @param end   end index, inclusive
      */
     public static void reverseFragment(int[] arr, int start, int end) {
-        assert start >= 0 && start <= arr.length: String.format("Start index (%s) must be in range [0, %s)", start, arr.length);
-        assert end >= 0 && end <= arr.length: String.format("End index (%s) must be in range [0, %s)", start, arr.length);
-        assert start <= end: String.format("Start index (%s) must be <= end (%s)", start, end);
+        assert start >= 0 && start <= arr.length : String.format("Start index (%s) must be in range [0, %s)", start, arr.length);
+        assert end >= 0 && end <= arr.length : String.format("End index (%s) must be in range [0, %s)", start, arr.length);
+        assert start <= end : String.format("Start index (%s) must be <= end (%s)", start, end);
         for (int i = start, j = end; i < j; i++, j--) {
             int temp = arr[i];
             arr[i] = arr[j];
@@ -37,14 +40,14 @@ public class ArrayUtil {
     /**
      * Reverse a fragment inside an array from start to end (inclusive)
      *
-     * @param arr Array to reverse
+     * @param arr   Array to reverse
      * @param start start index, inclusive
      * @param end   end index, inclusive
      */
     public static void reverseFragment(Object[] arr, int start, int end) {
-        assert start >= 0 && start <= arr.length: String.format("Start index (%s) must be in range [0, %s)", start, arr.length);
-        assert end >= 0 && end <= arr.length: String.format("End index (%s) must be in range [0, %s)", start, arr.length);
-        assert start <= end: String.format("Start index (%s) must be <= end (%s)", start, end);
+        assert start >= 0 && start <= arr.length : String.format("Start index (%s) must be in range [0, %s)", start, arr.length);
+        assert end >= 0 && end <= arr.length : String.format("End index (%s) must be in range [0, %s)", start, arr.length);
+        assert start <= end : String.format("Start index (%s) must be <= end (%s)", start, end);
         for (int i = start, j = end; i < j; i++, j--) {
             Object temp = arr[i];
             arr[i] = arr[j];
@@ -57,10 +60,9 @@ public class ArrayUtil {
      *
      * @param array Array to shuffle IN PLACE
      */
-    public static void shuffle(int[] array){
+    public static void shuffle(int[] array) {
         var rnd = RandomManager.getRandom();
-        for (int i = array.length - 1; i > 0; i--)
-        {
+        for (int i = array.length - 1; i > 0; i--) {
             int index = rnd.nextInt(i + 1);
             // Simple swap
             int a = array[index];
@@ -70,12 +72,11 @@ public class ArrayUtil {
     }
 
     /**
-     *
      * Swaps the two specified elements in the specified array.
      *
      * @param arr array
-     * @param i origin destination index
-     * @param j destination index
+     * @param i   origin destination index
+     * @param j   destination index
      */
     public static void swap(Object[] arr, int i, int j) {
         Object tmp = arr[i];
@@ -85,12 +86,11 @@ public class ArrayUtil {
 
 
     /**
-     *
      * Swaps the two specified elements in the specified array.
      *
      * @param arr array
-     * @param i origin destination index
-     * @param j destination index
+     * @param i   origin destination index
+     * @param j   destination index
      */
     public static void swap(int[] arr, int i, int j) {
         int tmp = arr[i];
@@ -99,12 +99,11 @@ public class ArrayUtil {
     }
 
     /**
-     *
      * Swaps the two specified elements in the specified array.
      *
      * @param arr array
-     * @param i origin destination index
-     * @param j destination index
+     * @param i   origin destination index
+     * @param j   destination index
      */
     public static void swap(double[] arr, int i, int j) {
         double tmp = arr[i];
@@ -113,12 +112,11 @@ public class ArrayUtil {
     }
 
     /**
-     *
      * Swaps the two specified elements in the specified array.
      *
      * @param arr array
-     * @param i origin destination index
-     * @param j destination index
+     * @param i   origin destination index
+     * @param j   destination index
      */
     public static void swap(long[] arr, int i, int j) {
         long tmp = arr[i];
@@ -133,7 +131,7 @@ public class ArrayUtil {
      * @param array Array to shuffle
      * @return shuffled array. Original array is not modified
      */
-    public static int[] copyAndshuffle(int[] array){
+    public static int[] copyAndshuffle(int[] array) {
         int[] copy = array.clone();
         shuffle(copy);
         return copy;
@@ -144,10 +142,9 @@ public class ArrayUtil {
      *
      * @param array Array to shuffle IN PLACE
      */
-    public static void shuffle(Object[] array){
+    public static void shuffle(Object[] array) {
         var rnd = RandomManager.getRandom();
-        for (int i = array.length - 1; i > 0; i--)
-        {
+        for (int i = array.length - 1; i > 0; i--) {
             int index = rnd.nextInt(i + 1);
             // Simple swap
             swap(array, index, i);
@@ -160,22 +157,22 @@ public class ArrayUtil {
      * Example: deleteAndInsert([a,b,c,d,e,f], 1, 4) = [a,c,d,e,b,f]
      * Example: deleteAndInsert([a,b,c,d,e,f], 5, 3) = [a,b,c,f,d,e]
      *
-     * @param array Array to modify
-     * @param origin index of element to be removed
+     * @param array       Array to modify
+     * @param origin      index of element to be removed
      * @param destination index where element will be inserted
-     * @param <T> Array type
+     * @param <T>         Array type
      * @return Modified Array
      */
-    public static <T> T[] deleteAndInsert(T[] array, int origin, int destination){
-        if(origin == destination) return array;
+    public static <T> T[] deleteAndInsert(T[] array, int origin, int destination) {
+        if (origin == destination) return array;
         T element = array[origin];
         int length;
-        if(origin < destination){
+        if (origin < destination) {
             length = destination - origin;
-            System.arraycopy(array, origin+1, array, origin,length);
+            System.arraycopy(array, origin + 1, array, origin, length);
         } else { // destination > origin
             length = origin - destination;
-            System.arraycopy(array, destination, array, destination+1,length);
+            System.arraycopy(array, destination, array, destination + 1, length);
         }
         array[destination] = element;
         return array;
@@ -186,21 +183,21 @@ public class ArrayUtil {
      * Example: deleteAndInsert([a,b,c,d,e,f], 0, 1) = [b,a,c,d,e,f]
      * Example: deleteAndInsert([a,b,c,d,e,f], 1, 4) = [a,c,d,e,b,f]
      *
-     * @param array Array to modify
-     * @param origin index of element to be removed
+     * @param array       Array to modify
+     * @param origin      index of element to be removed
      * @param destination index where element will be inserted
      * @return Modified Array
      */
-    public static int[] deleteAndInsert(int[] array, int origin, int destination){
-        if(origin == destination) return array;
+    public static int[] deleteAndInsert(int[] array, int origin, int destination) {
+        if (origin == destination) return array;
         int element = array[origin];
         int length;
-        if(origin < destination){
+        if (origin < destination) {
             length = destination - origin;
-            System.arraycopy(array, origin+1, array, origin,length);
+            System.arraycopy(array, origin + 1, array, origin, length);
         } else { // destination > origin
             length = origin - destination;
-            System.arraycopy(array, destination, array, destination+1,length);
+            System.arraycopy(array, destination, array, destination + 1, length);
         }
         array[destination] = element;
         return array;
@@ -212,21 +209,21 @@ public class ArrayUtil {
      * Example: deleteAndInsert([a,b,c,d,e,f], 0, 1) = [b,a,c,d,e,f]
      * Example: deleteAndInsert([a,b,c,d,e,f], 1, 4) = [a,c,d,e,b,f]
      *
-     * @param array Array to modify
-     * @param origin index of element to be removed
+     * @param array       Array to modify
+     * @param origin      index of element to be removed
      * @param destination index where element will be inserted
      * @return Modified Array
      */
-    public static long[] deleteAndInsert(long[] array, int origin, int destination){
-        if(origin == destination) return array;
+    public static long[] deleteAndInsert(long[] array, int origin, int destination) {
+        if (origin == destination) return array;
         long element = array[origin];
         int length;
-        if(origin < destination){
+        if (origin < destination) {
             length = destination - origin;
-            System.arraycopy(array, origin+1, array, origin,length);
+            System.arraycopy(array, origin + 1, array, origin, length);
         } else { // destination > origin
             length = origin - destination;
-            System.arraycopy(array, destination, array, destination+1,length);
+            System.arraycopy(array, destination, array, destination + 1, length);
         }
         array[destination] = element;
         return array;
@@ -236,12 +233,12 @@ public class ArrayUtil {
      * Insert element in given position. Elements to the right are shifted one position to the right.
      * Rightmost element is dropped.
      *
-     * @param arr Array to modify
+     * @param arr   Array to modify
      * @param index Position in which insert the element
      * @param value Element to insert
      */
-    public static void insert(int[] arr, int index, int value){
-        System.arraycopy(arr, index, arr, index+1, arr.length - index -1);
+    public static void insert(int[] arr, int index, int value) {
+        System.arraycopy(arr, index, arr, index + 1, arr.length - index - 1);
         arr[index] = value;
     }
 
@@ -249,12 +246,12 @@ public class ArrayUtil {
      * Insert element in given position. Elements to the right are shifted one position to the right.
      * Rightmost element is dropped.
      *
-     * @param arr Array to modify
+     * @param arr   Array to modify
      * @param index Position in which insert the element
      * @param value Element to insert
      */
-    public static void insert(long[] arr, int index, long value){
-        System.arraycopy(arr, index, arr, index+1, arr.length - index -1);
+    public static void insert(long[] arr, int index, long value) {
+        System.arraycopy(arr, index, arr, index + 1, arr.length - index - 1);
         arr[index] = value;
     }
 
@@ -262,12 +259,12 @@ public class ArrayUtil {
      * Insert element in given position. Elements to the right are shifted one position to the right.
      * Rightmost element is dropped.
      *
-     * @param arr Array to modify
+     * @param arr   Array to modify
      * @param index Position in which insert the element
      * @param value Element to insert
      */
-    public static void insert(double[] arr, int index, double value){
-        System.arraycopy(arr, index, arr, index+1, arr.length - index -1);
+    public static void insert(double[] arr, int index, double value) {
+        System.arraycopy(arr, index, arr, index + 1, arr.length - index - 1);
         arr[index] = value;
     }
 
@@ -275,13 +272,13 @@ public class ArrayUtil {
      * Insert element in given position. Elements to the right are shifted one position to the right.
      * Rightmost element is dropped.
      *
-     * @param arr Array to modify
+     * @param arr   Array to modify
      * @param index Position in which insert the element
      * @param value Element to insert
-     * @param <T> type
+     * @param <T>   type
      */
-    public static <T> void insert(T[] arr, int index, T value){
-        System.arraycopy(arr, index, arr, index+1, arr.length - index -1);
+    public static <T> void insert(T[] arr, int index, T value) {
+        System.arraycopy(arr, index, arr, index + 1, arr.length - index - 1);
         arr[index] = value;
     }
 
@@ -289,13 +286,13 @@ public class ArrayUtil {
      * Remove element at given index and shift elements to the left. Rightmost element is duplicated.
      * Example: remove([9,10,11,12], 1) → [9,11,12,12]
      *
-     * @param arr array to modify
+     * @param arr   array to modify
      * @param index index of element to delete
      * @return removed element
      */
-    public static int remove(int[] arr, int index){
+    public static int remove(int[] arr, int index) {
         int value = arr[index];
-        System.arraycopy(arr, index+1, arr, index, arr.length - index -1);
+        System.arraycopy(arr, index + 1, arr, index, arr.length - index - 1);
         return value;
     }
 
@@ -303,13 +300,13 @@ public class ArrayUtil {
      * Remove element at given index and shift elements to the left. Rightmost element is duplicated.
      * Example: remove([9,10,11,12], 1) → [9,11,12,12]
      *
-     * @param arr array to modify
+     * @param arr   array to modify
      * @param index index of element to delete
      * @return removed element
      */
-    public static long remove(long[] arr, int index){
+    public static long remove(long[] arr, int index) {
         long value = arr[index];
-        System.arraycopy(arr, index+1, arr, index, arr.length - index -1);
+        System.arraycopy(arr, index + 1, arr, index, arr.length - index - 1);
         return value;
     }
 
@@ -317,13 +314,13 @@ public class ArrayUtil {
      * Remove element at given index and shift elements to the left. Rightmost element is duplicated.
      * Example: remove([9,10,11,12], 1) → [9,11,12,12]
      *
-     * @param arr array to modify
+     * @param arr   array to modify
      * @param index index of element to delete
      * @return removed element
      */
-    public static double remove(double[] arr, int index){
+    public static double remove(double[] arr, int index) {
         double value = arr[index];
-        System.arraycopy(arr, index+1, arr, index, arr.length - index -1);
+        System.arraycopy(arr, index + 1, arr, index, arr.length - index - 1);
         return value;
     }
 
@@ -331,14 +328,14 @@ public class ArrayUtil {
      * Remove element at given index and shift elements to the left. Rightmost element is duplicated.
      * Example: remove([9,10,11,12], 1) → [9,11,12,12]
      *
-     * @param arr array to modify
+     * @param arr   array to modify
      * @param index index of element to delete
-     * @param <T> type
+     * @param <T>   type
      * @return removed element
      */
-    public static <T> T remove(T[] arr, int index){
+    public static <T> T remove(T[] arr, int index) {
         T value = arr[index];
-        System.arraycopy(arr, index+1, arr, index, arr.length - index -1);
+        System.arraycopy(arr, index + 1, arr, index, arr.length - index - 1);
         return value;
     }
 
@@ -348,14 +345,14 @@ public class ArrayUtil {
      * @param data array data
      * @return flattened array
      */
-    public static int[] flatten(int[][] data){
+    public static int[] flatten(int[][] data) {
         int size = 0;
-        for(var row: data){
+        for (var row : data) {
             size += row.length;
         }
         int left = 0;
         var result = new int[size];
-        for (var row: data) {
+        for (var row : data) {
             System.arraycopy(row, 0, result, left, row.length);
             left += row.length;
         }
@@ -368,14 +365,14 @@ public class ArrayUtil {
      * @param data array data
      * @return flattened array
      */
-    public static double[] flatten(double[][] data){
+    public static double[] flatten(double[][] data) {
         int size = 0;
-        for(var row: data){
+        for (var row : data) {
             size += row.length;
         }
         int left = 0;
         var result = new double[size];
-        for (var row: data) {
+        for (var row : data) {
             System.arraycopy(row, 0, result, left, row.length);
             left += row.length;
         }
@@ -388,14 +385,14 @@ public class ArrayUtil {
      * @param data array data
      * @return flattened array
      */
-    public static long[] flatten(long[][] data){
+    public static long[] flatten(long[][] data) {
         int size = 0;
-        for(var row: data){
+        for (var row : data) {
             size += row.length;
         }
         int left = 0;
         var result = new long[size];
-        for (var row: data) {
+        for (var row : data) {
             System.arraycopy(row, 0, result, left, row.length);
             left += row.length;
         }
@@ -408,7 +405,7 @@ public class ArrayUtil {
      * @param data array
      * @return number of non null elements
      */
-    public static int countNonNull(Object[] data){
+    public static int countNonNull(Object[] data) {
         return data.length - countNull(data);
     }
 
@@ -418,10 +415,10 @@ public class ArrayUtil {
      * @param data array
      * @return number of null elements
      */
-    public static int countNull(Object[] data){
+    public static int countNull(Object[] data) {
         int count = 0;
-        for(var d: data){
-            if(d == null){
+        for (var d : data) {
+            if (d == null) {
                 count++;
             }
         }
@@ -431,13 +428,13 @@ public class ArrayUtil {
     /**
      * Sum all elements in array
      *
-     * @throws java.lang.ArithmeticException if there is an overflow
      * @param data numbers to sum
      * @return sum of all numbers
+     * @throws java.lang.ArithmeticException if there is an overflow
      */
-    public static int sum(int[] data){
+    public static int sum(int[] data) {
         int sum = 0;
-        for(int i: data){
+        for (int i : data) {
             sum = Math.addExact(sum, i);
         }
         return sum;
@@ -449,9 +446,9 @@ public class ArrayUtil {
      * @param data numbers to sum
      * @return sum of all numbers
      */
-    public static double sum(double[] data){
+    public static double sum(double[] data) {
         int sum = 0;
-        for(double i: data){
+        for (double i : data) {
             sum += i;
         }
         return sum;
@@ -460,15 +457,133 @@ public class ArrayUtil {
     /**
      * Sum all elements in array
      *
-     * @throws java.lang.ArithmeticException if there is an overflow
      * @param data numbers to sum
      * @return sum of all numbers
+     * @throws java.lang.ArithmeticException if there is an overflow
      */
-    public static long sum(long[] data){
+    public static long sum(long[] data) {
         long sum = 0;
-        for(long i: data){
+        for (long i : data) {
             sum = Math.addExact(sum, i);
         }
         return sum;
+    }
+
+    /**
+     * Find the biggest value in the given array
+     * @param values array of double numbers. Both positive and negative infinity are valid values.
+     * @return biggest value in array
+     * @throws IllegalArgumentException if the array contains a NaN
+     */
+    public static double max(double[] values) {
+        if(values.length == 0){
+            throw new IllegalArgumentException("Empty array");
+        }
+        double max = Double.NEGATIVE_INFINITY;
+        for (int i = 0; i < values.length; i++) {
+            double d = values[i];
+            if (Double.isNaN(d)) {
+                throw new IllegalArgumentException("NaN at index " + i);
+            }
+            if (d > max) {
+                max = d;
+            }
+        }
+        return max;
+    }
+
+    /**
+     * Find the biggest value in the given array
+     * @param values array of integer numbers.
+     * @return biggest value in array
+     */
+    public static int max(int[] values) {
+        if(values.length == 0){
+            throw new IllegalArgumentException("Empty array");
+        }
+        int max = Integer.MIN_VALUE;
+        for (int d : values) {
+            if (d > max) {
+                max = d;
+            }
+        }
+        return max;
+    }
+
+    /**
+     * Find the biggest value in the given array
+     * @param values array of long numbers.
+     * @return biggest value in array
+     */
+    public static long max(long[] values) {
+        if(values.length == 0){
+            throw new IllegalArgumentException("Empty array");
+        }
+        long max = Long.MIN_VALUE;
+        for (long d : values) {
+            if (d > max) {
+                max = d;
+            }
+        }
+        return max;
+    }
+
+    /**
+     * Find the smallest value in the given array
+     * @param values array of double numbers. Both positive and negative infinity are valid values.
+     * @return smallest value in array
+     * @throws IllegalArgumentException if the array contains a NaN
+     */
+    public static double min(double[] values) {
+        if(values.length == 0){
+            throw new IllegalArgumentException("Empty array");
+        }
+        double min = Double.POSITIVE_INFINITY;
+        for (int i = 0; i < values.length; i++) {
+            double d = values[i];
+            if (Double.isNaN(d)) {
+                throw new IllegalArgumentException("NaN at index " + i);
+            }
+            if (d < min) {
+                min = d;
+            }
+        }
+        return min;
+    }
+
+    /**
+     * Find the smallest value in the given array
+     * @param values array of integer numbers.
+     * @return smallest value in array
+     */
+    public static int min(int[] values) {
+        if(values.length == 0){
+            throw new IllegalArgumentException("Empty array");
+        }
+        int min = Integer.MAX_VALUE;
+        for (int d : values) {
+            if (d < min) {
+                min = d;
+            }
+        }
+        return min;
+    }
+
+    /**
+     * Find the smallest value in the given array
+     * @param values array of long numbers.
+     * @return smallest value in array
+     */
+    public static long min(long[] values) {
+        if(values.length == 0){
+            throw new IllegalArgumentException("Empty array");
+        }
+        long min = Long.MAX_VALUE;
+        for (long d : values) {
+            if (d < min) {
+                min = d;
+            }
+        }
+        return min;
     }
 }