+++ common/src/test/java/es/urjc/etsii/grafo/util/ArrayUtilTest.java
@@ -6,10 +6,10 @@ import org.junit.jupiter.api.Test;
 
 import java.util.Arrays;
 
-public class ArrayUtilTest {
+class ArrayUtilTest {
 
     @Test
-    public void testInsertAndDeleteRightIntegersDistance1() {
+    void testInsertAndDeleteRightIntegersDistance1() {
         Integer[] original = new Integer[]{0, 1, 2, 3, 4, 5, 6, 7, 8, 9};
         Integer[] expected = new Integer[]{1, 0, 2, 3, 4, 5, 6, 7, 8, 9};
         ArrayUtil.deleteAndInsert(original, 0, 1);
@@ -17,7 +17,7 @@ public class ArrayUtilTest {
     }
 
     @Test
-    public void testInsertAndDeleteRightIntegersDistance4() {
+    void testInsertAndDeleteRightIntegersDistance4() {
         Integer[] original = new Integer[]{0, 1, 2, 3, 4, 5, 6, 7, 8, 9};
         Integer[] expected = new Integer[]{0, 2, 3, 4, 5, 1, 6, 7, 8, 9};
         ArrayUtil.deleteAndInsert(original, 1, 5);
@@ -25,7 +25,7 @@ public class ArrayUtilTest {
     }
 
     @Test
-    public void testInsertAndDeleteRightStrings() {
+    void testInsertAndDeleteRightStrings() {
         String[] original = new String[]{"a", "b", "c", "d", "e", "f", "g", "h", "i", "j", "k", "l"};
         String[] expected = new String[]{"a", "c", "d", "e", "f", "g", "h", "i", "j", "k", "b", "l"};
         ArrayUtil.deleteAndInsert(original, 1, 10);
@@ -33,7 +33,7 @@ public class ArrayUtilTest {
     }
 
     @Test
-    public void testInsertAndDeleteRightInts() {
+    void testInsertAndDeleteRightInts() {
         int[] original = new int[]{0, 1, 2, 3, 4, 5, 6, 7, 8, 9};
         int[] expected = new int[]{0, 2, 3, 4, 5, 6, 7, 8, 9, 1};
         ArrayUtil.deleteAndInsert(original, 1, 9);
@@ -41,7 +41,7 @@ public class ArrayUtilTest {
     }
 
     @Test
-    public void testInsertAndDeleteRightLongs() {
+    void testInsertAndDeleteRightLongs() {
         long[] original = new long[]{0L, 1L, 2L, 3L, 4L, 5L, 6L, 7L, 8L, 9L};
         long[] expected = new long[]{1L, 2L, 3L, 4L, 5L, 6L, 7L, 8L, 9L, 0L};
         ArrayUtil.deleteAndInsert(original, 0, 9);
@@ -50,7 +50,7 @@ public class ArrayUtilTest {
 
     // Same as above but for left insertion
     @Test
-    public void testInsertAndDeleteLeftIntegersDistance1() {
+    void testInsertAndDeleteLeftIntegersDistance1() {
         Integer[] original = new Integer[]{0, 1, 2, 3, 4, 5, 6, 7, 8, 9};
         Integer[] expected = new Integer[]{0, 1, 2, 3, 4, 5, 6, 7, 9, 8};
         ArrayUtil.deleteAndInsert(original, 9, 8);
@@ -58,7 +58,7 @@ public class ArrayUtilTest {
     }
 
     @Test
-    public void testInsertAndDeleteLeftIntegersDistance4() {
+    void testInsertAndDeleteLeftIntegersDistance4() {
         Integer[] original = new Integer[]{0, 1, 2, 3, 4, 5, 6, 7, 8, 9};
         Integer[] expected = new Integer[]{0, 1, 2, 3, 8, 4, 5, 6, 7, 9};
         ArrayUtil.deleteAndInsert(original, 8, 4);
@@ -66,7 +66,7 @@ public class ArrayUtilTest {
     }
 
     @Test
-    public void testInsertAndDeleteLeftStrings() {
+    void testInsertAndDeleteLeftStrings() {
         String[] original = new String[]{"a", "b", "c", "d", "e", "f", "g", "h", "i", "j", "k", "l"};
         String[] expected = new String[]{"a", "k", "b", "c", "d", "e", "f", "g", "h", "i", "j", "l"};
         ArrayUtil.deleteAndInsert(original, 10, 1);
@@ -74,7 +74,7 @@ public class ArrayUtilTest {
     }
 
     @Test
-    public void testInsertAndDeleteLeftInts() {
+    void testInsertAndDeleteLeftInts() {
         int[] original = new int[]{0, 1, 2, 3, 4, 5, 6, 7, 8, 9};
         int[] expected = new int[]{0, 9, 1, 2, 3, 4, 5, 6, 7, 8};
         ArrayUtil.deleteAndInsert(original, 9, 1);
@@ -82,7 +82,7 @@ public class ArrayUtilTest {
     }
 
     @Test
-    public void testInsertAndDeleteLeftLongs() {
+    void testInsertAndDeleteLeftLongs() {
         long[] original = new long[]{0L, 1L, 2L, 3L, 4L, 5L, 6L, 7L, 8L, 9L};
         long[] expected = new long[]{9L, 0L, 1L, 2L, 3L, 4L, 5L, 6L, 7L, 8L};
         ArrayUtil.deleteAndInsert(original, 9, 0);
@@ -90,7 +90,7 @@ public class ArrayUtilTest {
     }
 
     @Test
-    public void testUndo() {
+    void testUndo() {
         long[] original = new long[]{0L, 1L, 2L, 3L, 4L, 5L, 6L, 7L, 8L, 9L};
         long[] copy = Arrays.copyOf(original, original.length);
         ArrayUtil.deleteAndInsert(original, 3, 8);
@@ -99,98 +99,98 @@ public class ArrayUtilTest {
     }
 
     @Test
-    public void testIntInsertArray() {
+    void testIntInsertArray() {
         int[] orig = {0, 1, 2, 3, 4};
         ArrayUtil.insert(orig, 0, 9);
-        Assertions.assertArrayEquals(orig, new int[]{9, 0, 1, 2, 3});
+        Assertions.assertArrayEquals(new int[]{9, 0, 1, 2, 3}, orig);
 
         ArrayUtil.insert(orig, 4, 99);
-        Assertions.assertArrayEquals(orig, new int[]{9, 0, 1, 2, 99});
+        Assertions.assertArrayEquals(new int[]{9, 0, 1, 2, 99}, orig);
     }
 
     @Test
-    public void testLongInsertArray() {
+    void testLongInsertArray() {
         long[] orig = {0, 1, 2, 3, 4};
         ArrayUtil.insert(orig, 0, 9);
-        Assertions.assertArrayEquals(orig, new long[]{9, 0, 1, 2, 3});
+        Assertions.assertArrayEquals(new long[]{9, 0, 1, 2, 3}, orig);
 
         ArrayUtil.insert(orig, 4, 99);
-        Assertions.assertArrayEquals(orig, new long[]{9, 0, 1, 2, 99});
+        Assertions.assertArrayEquals(new long[]{9, 0, 1, 2, 99}, orig);
     }
 
     @Test
-    public void testDoubleInsertArray() {
+    void testDoubleInsertArray() {
         double[] orig = {0d, 1d, 2d, 3d, 4d};
         ArrayUtil.insert(orig, 0, 9);
-        Assertions.assertArrayEquals(orig, new double[]{9d, 0d, 1d, 2d, 3d});
+        Assertions.assertArrayEquals(new double[]{9d, 0d, 1d, 2d, 3d}, orig);
 
         ArrayUtil.insert(orig, 4, 99);
-        Assertions.assertArrayEquals(orig, new double[]{9d, 0d, 1d, 2d, 99d});
+        Assertions.assertArrayEquals(new double[]{9d, 0d, 1d, 2d, 99d}, orig);
     }
 
     @Test
-    public void testTInsertArray() {
+    void testTInsertArray() {
         Integer[] orig = {0, 1, 2, 3, 4};
         ArrayUtil.insert(orig, 0, 9);
-        Assertions.assertArrayEquals(orig, new Integer[]{9, 0, 1, 2, 3});
+        Assertions.assertArrayEquals(new Integer[]{9, 0, 1, 2, 3}, orig);
 
         ArrayUtil.insert(orig, 4, 99);
-        Assertions.assertArrayEquals(orig, new Integer[]{9, 0, 1, 2, 99});
+        Assertions.assertArrayEquals(new Integer[]{9, 0, 1, 2, 99}, orig);
     }
 
     @Test
-    public void testIntDeleteArray() {
+    void testIntDeleteArray() {
         int[] orig = {0, 1, 2, 3, 4};
         ArrayUtil.remove(orig, 0);
-        Assertions.assertArrayEquals(orig, new int[]{1, 2, 3, 4, 4});
+        Assertions.assertArrayEquals(new int[]{1, 2, 3, 4, 4}, orig);
 
         ArrayUtil.remove(orig, 4);
-        Assertions.assertArrayEquals(orig, new int[]{1, 2, 3, 4, 4});
+        Assertions.assertArrayEquals(new int[]{1, 2, 3, 4, 4}, orig);
 
         ArrayUtil.remove(orig, 1);
-        Assertions.assertArrayEquals(orig, new int[]{1, 3, 4, 4, 4});
+        Assertions.assertArrayEquals(new int[]{1, 3, 4, 4, 4}, orig);
     }
 
     @Test
-    public void testLongDeleteArray() {
+    void testLongDeleteArray() {
         long[] orig = {0, 1, 2, 3, 4};
         ArrayUtil.remove(orig, 0);
-        Assertions.assertArrayEquals(orig, new long[]{1, 2, 3, 4, 4});
+        Assertions.assertArrayEquals(new long[]{1, 2, 3, 4, 4}, orig);
 
         ArrayUtil.remove(orig, 4);
-        Assertions.assertArrayEquals(orig, new long[]{1, 2, 3, 4, 4});
+        Assertions.assertArrayEquals(new long[]{1, 2, 3, 4, 4}, orig);
 
         ArrayUtil.remove(orig, 1);
-        Assertions.assertArrayEquals(orig, new long[]{1, 3, 4, 4, 4});
+        Assertions.assertArrayEquals(new long[]{1, 3, 4, 4, 4}, orig);
     }
 
     @Test
-    public void testDoubleDeleteArray() {
+    void testDoubleDeleteArray() {
         double[] orig = {0, 1, 2, 3, 4};
         ArrayUtil.remove(orig, 0);
-        Assertions.assertArrayEquals(orig, new double[]{1, 2, 3, 4, 4});
+        Assertions.assertArrayEquals(new double[]{1, 2, 3, 4, 4}, orig);
 
         ArrayUtil.remove(orig, 4);
-        Assertions.assertArrayEquals(orig, new double[]{1, 2, 3, 4, 4});
+        Assertions.assertArrayEquals(new double[]{1, 2, 3, 4, 4}, orig);
 
         ArrayUtil.remove(orig, 1);
-        Assertions.assertArrayEquals(orig, new double[]{1, 3, 4, 4, 4});
+        Assertions.assertArrayEquals(new double[]{1, 3, 4, 4, 4}, orig);
     }
     @Test
-    public void testTDeleteArray() {
+    void testTDeleteArray() {
         Integer[] orig = {0, 1, 2, 3, 4};
         ArrayUtil.remove(orig, 0);
-        Assertions.assertArrayEquals(orig, new Integer[]{1, 2, 3, 4, 4});
+        Assertions.assertArrayEquals(new Integer[]{1, 2, 3, 4, 4}, orig);
 
         ArrayUtil.remove(orig, 4);
-        Assertions.assertArrayEquals(orig, new Integer[]{1, 2, 3, 4, 4});
+        Assertions.assertArrayEquals(new Integer[]{1, 2, 3, 4, 4}, orig);
 
         ArrayUtil.remove(orig, 1);
-        Assertions.assertArrayEquals(orig, new Integer[]{1, 3, 4, 4, 4});
+        Assertions.assertArrayEquals(new Integer[]{1, 3, 4, 4, 4}, orig);
     }
 
     @Test
-    public void testFlattenInt() {
+    void testFlattenInt() {
         int[][] original = {
                 {0,1,2},
                 {3},
@@ -207,7 +207,7 @@ public class ArrayUtilTest {
     }
 
     @Test
-    public void testFlattenLong() {
+    void testFlattenLong() {
         long[][] original = {
                 {0,1,2},
                 {3},
@@ -224,7 +224,7 @@ public class ArrayUtilTest {
     }
 
     @Test
-    public void testFlattenDouble() {
+    void testFlattenDouble() {
         double[][] original = {
                 {0,1,2},
                 {3},
@@ -241,7 +241,7 @@ public class ArrayUtilTest {
     }
 
     @Test
-    public void countNulls(){
+    void countNulls(){
         Object[] empty = new Integer[]{};
         Assertions.assertEquals(0, ArrayUtil.countNull(empty));
         Assertions.assertEquals(0, ArrayUtil.countNonNull(empty));
@@ -252,27 +252,27 @@ public class ArrayUtilTest {
     }
 
     @Test
-    public void testSumInts(){
+    void testSumInts(){
         Assertions.assertThrows(ArithmeticException.class, () -> ArrayUtil.sum(new int[]{Integer.MAX_VALUE, 1}));
         Assertions.assertEquals(5, ArrayUtil.sum(new int[]{2, 3}));
         Assertions.assertEquals(-1, ArrayUtil.sum(new int[]{-2, 1}));
     }
 
     @Test
-    public void testSumLongs(){
+    void testSumLongs(){
         Assertions.assertThrows(ArithmeticException.class, () -> ArrayUtil.sum(new long[]{Long.MAX_VALUE, 1}));
         Assertions.assertEquals(5, ArrayUtil.sum(new long[]{2, 3}));
         Assertions.assertEquals(-1, ArrayUtil.sum(new long[]{-2, 1}));
     }
 
     @Test
-    public void testSumDouble(){
+    void testSumDouble(){
         Assertions.assertEquals(5D, ArrayUtil.sum(new double[]{2, 3}), 0.001);
         Assertions.assertEquals(-1D, ArrayUtil.sum(new double[]{-2, 1}), 0.001);
     }
 
     @Test
-    public void testSwapObjects(){
+    void testSwapObjects(){
         TestInstance first = new TestInstance("first");
         TestInstance second = new TestInstance("second");
         TestInstance third = new TestInstance("third");
@@ -287,30 +287,74 @@ public class ArrayUtilTest {
     }
 
     @Test
-    public void testSwapInts(){
+    void testSwapInts(){
         int[] data = new int[]{0,1,2,3,4,5};
         ArrayUtil.swap(data, 0, 2);
-        Assertions.assertArrayEquals(data, new int[]{2,1,0,3,4,5});
+        Assertions.assertArrayEquals(new int[]{2, 1, 0, 3, 4, 5}, data);
         ArrayUtil.swap(data, 4, 5);
-        Assertions.assertArrayEquals(data, new int[]{2,1,0,3,5,4});
+        Assertions.assertArrayEquals(new int[]{2, 1, 0, 3, 5, 4}, data);
     }
 
     @Test
-    public void testSwapDoubles(){
+    void testSwapDoubles(){
         double[] data = new double[]{0,1,2,3,4,5};
         ArrayUtil.swap(data, 0, 2);
-        Assertions.assertArrayEquals(data, new double[]{2,1,0,3,4,5});
+        Assertions.assertArrayEquals(new double[]{2, 1, 0, 3, 4, 5}, data);
         ArrayUtil.swap(data, 4, 5);
-        Assertions.assertArrayEquals(data, new double[]{2,1,0,3,5,4});
+        Assertions.assertArrayEquals(new double[]{2, 1, 0, 3, 5, 4}, data);
     }
 
     @Test
-    public void testSwapLongs(){
+    void testSwapLongs(){
         long[] data = new long[]{0,1,2,3,4,5};
         ArrayUtil.swap(data, 0, 2);
-        Assertions.assertArrayEquals(data, new long[]{2,1,0,3,4,5});
+        Assertions.assertArrayEquals(new long[]{2, 1, 0, 3, 4, 5}, data);
         ArrayUtil.swap(data, 4, 5);
-        Assertions.assertArrayEquals(data, new long[]{2,1,0,3,5,4});
+        Assertions.assertArrayEquals(new long[]{2, 1, 0, 3, 5, 4}, data);
     }
 
+    @Test
+    void testMaxInt(){
+        Assertions.assertThrows(IllegalArgumentException.class, () -> ArrayUtil.max(new int[]{}));
+        Assertions.assertEquals(0, ArrayUtil.max(new int[]{-1, 0, -4123}));
+        Assertions.assertEquals(0, ArrayUtil.max(new int[]{0}));
+    }
+
+    @Test
+    void testMaxLong(){
+        Assertions.assertThrows(IllegalArgumentException.class, () -> ArrayUtil.max(new long[]{}));
+        Assertions.assertEquals(0L, ArrayUtil.max(new long[]{-1, 0, -4123}));
+        Assertions.assertEquals(0L, ArrayUtil.max(new long[]{0}));
+    }
+
+    @Test
+    void testMaxDouble(){
+        Assertions.assertThrows(IllegalArgumentException.class, () -> ArrayUtil.max(new double[]{}));
+        Assertions.assertThrows(IllegalArgumentException.class, () -> ArrayUtil.max(new double[]{1,4,Double.NaN,8}));
+        Assertions.assertEquals(0, ArrayUtil.max(new double[]{-1, 0, -4123}));
+        Assertions.assertEquals(0, ArrayUtil.max(new double[]{0}));
+    }
+
+    @Test
+    void testMinInt(){
+        Assertions.assertThrows(IllegalArgumentException.class, () -> ArrayUtil.min(new int[]{}));
+        Assertions.assertEquals(-4123, ArrayUtil.min(new int[]{-1, 0, -4123}));
+        Assertions.assertEquals(0, ArrayUtil.min(new int[]{0}));
+    }
+
+    @Test
+    void testMinLong(){
+        Assertions.assertThrows(IllegalArgumentException.class, () -> ArrayUtil.min(new long[]{}));
+        Assertions.assertEquals(-4123L, ArrayUtil.min(new long[]{-1, 0, -4123}));
+        Assertions.assertEquals(0L, ArrayUtil.min(new long[]{0}));
+    }
+
+    @Test
+    void testMinDouble(){
+        Assertions.assertThrows(IllegalArgumentException.class, () -> ArrayUtil.min(new double[]{}));
+        Assertions.assertThrows(IllegalArgumentException.class, () -> ArrayUtil.min(new double[]{1,4,Double.NaN,8}));
+        Assertions.assertEquals(-4123D, ArrayUtil.min(new double[]{-1, 0, -4123}));
+        Assertions.assertEquals(0D, ArrayUtil.min(new double[]{0}));
+    }
+    
 }